# rpl-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BackEnd RPL - By Dewa Permana",
        "description": "Swagger Documentation BackEnd Web RPL - Created by Dewa Permana",
        "contact": {
            "email": "dewaprmanaptr@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "operationId": "c64fdb270498e81061530febf0dec4be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Dewa Permana"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "dewapermana@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|sometoken"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Selamat Datang User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid Credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout User",
                "description": "Logout user dengan menghapus token Sanctum yang aktif",
                "operationId": "e95cf528bb67fc0f040b55ef5a5aab8f",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Token tidak valid atau tidak dikirim)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari login (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}